% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstatErr-package.R
\docType{package}
\name{bstatErr}
\alias{bstatErr}
\title{bstatErr: Error-checking utilities for R}
\description{
Provides helper functions for type checking and validation.
}
\section{Functions}{

• check_logical(): validate logical inputs.
• check_string(): validate string inputs.
}

\author{
\strong{Maintainer}: Bert van der Werf \email{bert.vanderwerf@auckland.ac.nz} (\href{https://orcid.org/0000-0003-3072-5937}{ORCID})

}
