% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstatErr-package.R
\docType{package}
\name{bstatErr-package}
\alias{bstatErr-package}
\title{bstatErr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Bert van der Werf \email{bert.vanderwerf@auckland.ac.nz} (\href{https://orcid.org/0000-0003-3072-5937}{ORCID})

}
\keyword{internal}
