% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_frame.R
\name{check_data_frame}
\alias{check_data_frame}
\title{Check a data.frame for typical errors}
\usage{
check_data_frame(data_frame_arg, allow_null = FALSE, allow_zero_length = FALSE)
}
\arguments{
\item{allow_null}{Logical. If TRUE, NULL is accepted as a valid value. Default is FALSE.}

\item{allow_zero_length}{Logical. If TRUE, a data frame with zero number of rows is accepted as a valid value. Default is FALSE.}

\item{data.frame_arg}{data.frame to check.}
}
\value{
Invisibly returns \code{data_frame_arg} if all checks pass; otherwise throws an error.
}
\description{
Validates a single data.frame argument for type, length, NA and NULL data.frame value.
}
\examples{
check_data.frame(data.frame("abc",c(1,2,3)))
check_data.frame(NULL, allow_null = TRUE)
}
