% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function.R
\name{check_function}
\alias{check_function}
\title{Check a function for typical errors}
\usage{
check_function(function_arg, allow_null = FALSE)
}
\arguments{
\item{function_arg}{function to check.}

\item{allow_null}{Logical. If TRUE, NULL is accepted as a valid value. Default is FALSE.}
}
\value{
Invisibly returns \code{function_arg} if all checks pass; otherwise throws an error.
}
\description{
Validates a single function argument for type and NULL function.
}
\examples{
check_function(mean)
check_function(NULL, allow_null = TRUE)
}
