% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logical.R
\name{check_logical}
\alias{check_logical}
\title{Check a logical for typical errors}
\usage{
check_logical(logical_arg, allow_null = FALSE, allow_na = FALSE)
}
\arguments{
\item{logical_arg}{logical to check.}

\item{allow_null}{Logical. If TRUE, NULL is accepted as a valid value. Default is FALSE.}

\item{allow_na}{Logical. If TRUE, NA logicals are allowed. Default is FALSE.}
}
\value{
Invisibly returns \code{logical_arg} if all checks pass; otherwise throws an error.
}
\description{
Validates a single logical argument for type, length, NA and NULL logical value.
}
\examples{
check_logical(TRUE)
check_logical(NULL, allow_null = TRUE)
}
