% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_numeric}
\alias{check_numeric}
\title{Check a numeric for typical errors}
\usage{
check_numeric(numeric_arg, allow_null = FALSE, allow_na = FALSE)
}
\arguments{
\item{numeric_arg}{numeric to check.}

\item{allow_null}{Logical. If TRUE, NULL is accepted as a valid value. Default is FALSE.}

\item{allow_na}{Logical. If TRUE, NA numerics are allowed. Default is FALSE.}
}
\value{
Returns \code{numeric_arg} if all checks pass; otherwise throws an error.
}
\description{
Validates a single numeric argument for type, length, NA and NULL numeric value.
}
\examples{
check_numeric(TRUE)
check_numeric(NULL, allow_null = TRUE)
}
