% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric_vector.R
\name{check_numeric_vector}
\alias{check_numeric_vector}
\title{Check a numeric for typical errors}
\usage{
check_numeric_vector(
  numeric_arg,
  allow_null = FALSE,
  allow_na = FALSE,
  allow_inf = FALSE,
  allow_zero_length = FALSE
)
}
\arguments{
\item{numeric_arg}{numeric to check.}

\item{allow_null}{Logical. If TRUE, NULL is accepted as a valid value. Default is FALSE.}

\item{allow_na}{Logical. If TRUE, NA numerics are allowed. Default is FALSE.}

\item{allow_inf}{Logical. If TRUE, -Inf or Inf numerics are allowed. Default is FALSE.}

\item{allow_zero_length}{Logical. If TRUE, zero length vectors are allowed. Default is FALSE}
}
\value{
Invisibly returns \code{numeric_arg} if all checks pass; otherwise throws an error.
}
\description{
Validates a single numeric argument for type, length, NA and NULL numeric value.
}
\examples{
check_numeric_vector(TRUE)
check_numeric_vector(NULL, allow_null = TRUE)
}
